Try it with only the CLOCK_TIME_NONE definition in OS specific files
